完成了

+++ proto/model.go
@@ -2,19 +2,19 @@ package proto
 
 type GlobalWeight struct {
 	Iteration    int       `json:"iteration"`
-	WeightVector []float64 `json:"weight_vector"`
+	WeightVector []float32 `json:"weight_vector"`
 	Aggregator   []byte    `json:"aggregator"`
-	TotalAcc     float64   `json:"total_acc"`
+	TotalAcc     float32   `json:"total_acc"`
 }
 
 type LocalityWeight struct {
 	Iteration          int                        `json:"iteration"`
-	WeightVector       []float64                  `json:"weight_vector"`
+	WeightVector       []float32                  `json:"weight_vector"`
 	Trainer            []byte                     `json:"trainer"`
 	ValidatorSelection *SelectionResult           `json:"validator_selection"`
 	Losses             []*Envelope[*ValidateLoss] `json:"losses"`
-	Acc                float64                    `json:"acc"`
-	Loss               float64                    `json:"loss"`
+	Acc                float32                    `json:"acc"`
+	Loss               float32                    `json:"loss"`
 	DataNum            int                        `json:"data_num"`
 }
 
@@ -23,6 +23,6 @@ type ValidateLoss struct {
 	Trainer   []byte  `json:"trainer"`
 	Validator []byte  `json:"validator"`
 	ModelHash []byte  `json:"model_hash"`
-	Loss      float64 `json:"loss"`
-	Acc       float64 `json:"acc"`
+	Loss      float32 `json:"loss"`
+	Acc       float32 `json:"acc"`
 }
